refactor(THU-683): extract VAD endpointer; fix barge-in dead-air; test the gate
Addresses the thunder-deep-review findings on the barge-in change:

- Blocking: barge-in fired at bargeInFrames(4) < minSpeechFrames(8), so a 4–7
  frame speech blip could cut the assistant off and then be too short to commit —
  dead air. Barge-in now fires at minSpeechFrames, the same gate that guarantees a
  commit, so an interrupt always yields a replacement turn.
- Convention: the endpointing state machine is split into ./endpointer (pure, no
  mic) and unit-tested directly (blip → misfire, no barge-in; sustained → one
  barge-in + commit) rather than only via a session-level stub.
- Nit: note why the default TTS voice is `ryan`.
