+++ src/voice/audio/vad.ts
@@ -16,23 +16,15 @@
  * `getUserMedia` runs with echo cancellation on. For barge-in the session keeps
  * the gate running *during* playback, so browser AEC is the only thing stopping
  * the assistant's own audio from self-triggering — `onSpeechStart` deliberately
- * waits for `bargeInFrames` of sustained speech (not the first frame) to stay
- * robust against residual echo.
+ * waits for sustained speech (see the endpointer) to stay robust against residual
+ * echo. This file owns only the mic/worklet plumbing; the endpointing state
+ * machine lives in `./endpointer` so it can be unit-tested without a mic.
  */
-import { concatFrames } from '@/voice/engine/audio-engine'
 import { MediaDevicesUnavailableError } from '@/voice/voice-error'
+import { type EndpointerHandlers, createEndpointer } from './endpointer'
 
-export type VadHandlers = {
-  /** Sustained speech detected (~130 ms) — the barge-in trigger. Fires once per
-   *  utterance, ahead of `onUtterance`, so the session can cut off the assistant. */
-  onSpeechStart?: () => void
-  /** A committed utterance (mono Float32Array @ 16 kHz) ready for transcription. */
-  onUtterance: (audio: Float32Array) => void
-  /** Speech started but was too short to be a real utterance. */
-  onMisfire?: () => void
-  /** Per-frame mic RMS [0,1] while listening — drives the live waveform. */
-  onLevel?: (rms: number) => void
-}
+/** Handlers for the mic gate — same shape the endpointer consumes. */
+export type VadHandlers = EndpointerHandlers
 
 export type VadGate = {
   start: () => Promise<void>
@@ -43,98 +35,16 @@ export type VadGate = {
   setListening: (value: boolean) => void
 }
 
-/** Frames are 512 samples (~32 ms) — set by the capture worklet. */
-const speechRmsThreshold = 0.015 // normalized [-1,1] amplitude; tune for the mic
-const minSpeechFrames = 8 // ~256 ms — shorter is a misfire
-// Barge-in fires after this many *sustained* speech frames (~130 ms) instead of
-// on the first frame, so a brief residual-echo blip while the assistant is
-// speaking can't cut it off. Lower = snappier interrupts but more echo-driven
-// false triggers; raise this (or `speechRmsThreshold`) if the assistant keeps
-// interrupting itself. Kept below `minSpeechFrames` so barge-in leads the commit.
-const bargeInFrames = 4
-// ~1.4 s of trailing silence ends the turn. Long enough to think mid-sentence
-// without it firing on a natural pause; a streaming STT with semantic
-// endpointing would let us shorten this later.
-const endSilenceFrames = 45
-const prerollFrames = 4 // keep a little audio before onset so we don't clip it
-
 const aecConstraints: MediaStreamConstraints = {
   audio: { echoCancellation: true, noiseSuppression: true, autoGainControl: true },
-}
-
-const rms = (frame: Float32Array): number => {
-  let sum = 0
-  for (let i = 0; i < frame.length; i++) {
-    sum += frame[i] * frame[i]
-  }
-  return Math.sqrt(sum / frame.length)
 }
 
 export const createVadGate = (handlers: VadHandlers): VadGate => {
   let stream: MediaStream | null = null
   let ctx: AudioContext | null = null
   let node: AudioWorkletNode | null = null
-
   let listening = true
-  let speaking = false
-  let collected: Float32Array[] = []
-  let preroll: Float32Array[] = []
-  let silenceRun = 0
-  let speechFrames = 0 // frames actually above threshold (excludes trailing silence)
-  let bargeInFired = false // barge-in signalled once for the current utterance
-
-  const reset = () => {
-    speaking = false
-    collected = []
-    silenceRun = 0
-    speechFrames = 0
-    bargeInFired = false
-  }
-
-  const endUtterance = () => {
-    const frames = collected
-    const speech = speechFrames
-    reset()
-    // Gate on real speech only — `collected` also holds preroll + the ~1.4 s of
-    // trailing silence, so counting frames.length would never detect a misfire.
-    if (speech < minSpeechFrames) {
-      handlers.onMisfire?.()
-      return
-    }
-    handlers.onUtterance(concatFrames(frames))
-  }
-
-  const processFrame = (frame: Float32Array) => {
-    const level = rms(frame)
-    handlers.onLevel?.(level)
-    if (level >= speechRmsThreshold) {
-      if (!speaking) {
-        speaking = true
-        collected = [...preroll]
-      }
-      collected.push(frame)
-      silenceRun = 0
-      speechFrames++
-      // Confirmed speech (not a one-frame blip) — signal a possible barge-in
-      // exactly once. The utterance keeps accumulating and still commits via
-      // onUtterance, so the audio that triggered the barge-in isn't lost.
-      if (!bargeInFired && speechFrames >= bargeInFrames) {
-        bargeInFired = true
-        handlers.onSpeechStart?.()
-      }
-    } else if (speaking) {
-      collected.push(frame)
-      silenceRun++
-      if (silenceRun >= endSilenceFrames) {
-        endUtterance()
-      }
-    } else {
-      preroll.push(frame)
-      if (preroll.length > prerollFrames) {
-        preroll.shift()
-      }
-    }
-  }
+  const endpointer = createEndpointer(handlers)
 
   const start = async () => {
     // WKWebView hides `navigator.mediaDevices` outside a secure context (a Tauri
@@ -158,7 +68,7 @@ export const createVadGate = (handlers: VadHandlers): VadGate => {
     node.connect(ctx.destination) // worklet has no output; keeps the graph pulling
     node.port.onmessage = (event: MessageEvent<Float32Array>) => {
       if (listening) {
-        processFrame(event.data)
+        endpointer.processFrame(event.data)
       }
     }
   }
@@ -183,15 +93,13 @@ export const createVadGate = (handlers: VadHandlers): VadGate => {
     stream = null
     await ctx?.close()
     ctx = null
-    reset()
-    preroll = []
+    endpointer.clear()
   }
 
   const setListening = (value: boolean) => {
     listening = value
     if (!value) {
-      reset()
-      preroll = []
+      endpointer.clear()
     }
   }
